원을 그리는데까지는 됐는데, 포인트를 찍는 건 아직 안 되고 있음.

+++ projectile/ViewController.swift
@@ -24,7 +24,7 @@ class ViewController: UIViewController {
     @IBAction func sliderValue(_ sender: UISlider) {
         let sliderCurrentValue = Int(sender.value)
         
-        circleview.currentangle = sliderCurrentValue
+        circleview.currentangle = CGFloat(sliderCurrentValue)
         
         currentAngleLabel.text = "\(sliderCurrentValue)"
         setNeedsFocusUpdate()
